Fix tests

+++ code/TrackDb.UnitTest/DbTests/BlockSerializationTest.cs
@@ -22,9 +22,9 @@ namespace TrackDb.UnitTest.DbTests
             var blockBuilder = new BlockBuilder(schema);
             var buffer = new byte[5000];
 
-            blockBuilder.AppendRecord(["Alice", 12, 1]);
-            blockBuilder.AppendRecord(["Bob", 42, 2]);
-            blockBuilder.AppendRecord(["Carl", 5, 3]);
+            blockBuilder.AppendRecord(["Alice", 12, (long)1]);
+            blockBuilder.AppendRecord(["Bob", 42, (long)2]);
+            blockBuilder.AppendRecord(["Carl", 5, (long)3]);
 
             var stats = blockBuilder.Serialize(buffer);
             var block = ReadOnlyBlock.Load(buffer, schema);
